Fix unintentional delay of ASTs on MP mips64
Make the interrupt and trap return paths check for ASTs with interrupts
disabled. If there is no pending request, interrupts are kept disabled
until return to userspace. This fixes delayed processing of an AST
request that arrives from another CPU after the AST flag has been
checked but before the CPU has finished the return path. The IPI that
follows the setting of the AST flag forces kernel re-entry and recheck
of AST.

Noticed by guenther@