STP works fine on etherip(4), it was gif(4) that lacked a MAC address.
getting etherip(4) to talk to a switch so i could verify this was
pretty fun.

remi@ pointed this bit of the manpage out somewhere, but it didn't
seem right anymore since etherip(4) was split out of gif(4). these
days if we have an interface that can be added to a bridge, we try
to make it look like an actual ethernet interface, which includes
having a mac address. we don't really have interfaces that change
modes and have these edge cases anymore.