em: Fix potential endless loop
If the NIC is in some error state (seen on a i219LM when em_read_phy_reg_ex()
returns at "MDI Error"), it can happen that we loop endlessly because the loop
variable is modified again somewhere down in the call stack. Use a separate
variable to make the attach fail with "Hardware Initialization Failed" instead
of hanging.

yes deraadt@