back out the use of ifiq pressure, and go back to using a packet count.
the pressure thresholds were too low in a lot of situations, and
still produced hard to understand interactions at high thresholds.
until we understand the numbers better, and for release, we're going
back counting the length of the per interface input queues.

this was originally based on a report of bad tcp performance with
em(4) by mlarkin, but is very convincingly demonstrated by a bunch
of work procter@ has been doing. deraadt@ is keen on the pressure
backout so he can cut a release.