Better error handling for bwfm(4) connection attempts.  When we fail to connect, e.g. due to a timeout, we will switch the state to SCAN. Unfortunately this skips clearing the active channel set, which means that on a scan on all bands only the nodes on the active channel set, which is defined by whatever node we tried to connect to, are allowed and all other APs are ignored.  Fix this by properly calling begin_scan().  When we fail to connect and start a scan, make sure to let the chip know that we don't want to associate anymore.
Another issue existed when we interrupt a scan, for instance by
setting a new nwid or wpakey.  In this case we didn't abort the
scan and started a new scan while the old one as still active.
This could lead to a SCAN -> SCAN transition loop.

Remove the "set ssid" event, since this would be an event in
addition to a failed auth/assoc event, which would make us try
to handle failure twice.

Discussed with and ok stsp@