Add ogx(4), a driver for the OCTEON III network processor.
This network driver covers higher-end models of the OCTEON III family.
They have a modified design whose interface is not compatible with
the lower-end models or earlier chip generations.

The code is still a work in progress. However, it is capable enough
to make the SGMII port functional on the CN7360.

No objection from deraadt@